extract WP collection methods into helper class
